Finalise DMGenerationChestEvent
Tested with this little plugin :
https://dl.dropbox.com/u/60299638/ChestDiamond.jar

This plugin from me ,change all of Chest content to a single Diamond
(that's an example)

here is the source from my listener :

```java
@EventHandler (priority = EventPriority.NORMAL)
public void onChestGenerate (DMGenerationChestEvent event) {
event.setCancelled(true);
Random random = event.getRandom();
Chest chest = (Chest) event.getBlock().getState();
List<ItemStack> itemStack = event.getContents();
itemStack.clear();
itemStack.add(new ItemStack(Material.DIAMOND, 1));
event.addItemsToChest(random, chest, itemStack);
}
```

Xephi
