+++ src/com/timvisee/DungeonMaze/event/generation/DMGenerationChestEvent.java
@@ -1,9 +1,11 @@
 package com.timvisee.DungeonMaze.event.generation;
 
 import java.util.List;
+import java.util.Random;
 
 import org.bukkit.World;
 import org.bukkit.block.Block;
+import org.bukkit.block.Chest;
 import org.bukkit.inventory.ItemStack;
 
 import com.timvisee.DungeonMaze.event.EventHandler.DMEventHandler;
@@ -12,10 +14,12 @@ public class DMGenerationChestEvent extends DMEventHandler {
 	
 	private Block b;
 	private List<ItemStack> is;
+	private Random random;
 	
-	public DMGenerationChestEvent(Block b, List<ItemStack> is) {
+	public DMGenerationChestEvent(Block b, Random random, List<ItemStack> is) {
 		this.b = b;
 		this.is = is;
+		this.random = random;
 	}
 	
 	public Block getBlock() {
@@ -24,6 +28,10 @@ public class DMGenerationChestEvent extends DMEventHandler {
 	
 	public List<ItemStack> getContents() {
 		return this.is;
+	}
+	
+	public Random getRandom() {
+		return this.random;
 	}
 	
 	public void setContents(List<ItemStack> is) {
@@ -35,4 +43,12 @@ public class DMGenerationChestEvent extends DMEventHandler {
 		return this.b.getWorld();
 	}
 
+	public void addItemsToChest(Random random, Chest chest, List<ItemStack> newContents) {
+		// Add new content to a chest
+		chest.getInventory().clear();
+		for (int i = 0; i < newContents.size(); i++) {
+			chest.getInventory().setItem(random.nextInt(chest.getInventory().getSize()), newContents.get(i));
+		}
+		chest.update();
+	}
 }
